chore: add initial project setup
